Add warning for non real robots

+++ robot/install.bash
@@ -1,6 +1,17 @@
 #! /usr/bin/env bash
 [[ "$CI" == "true" ]] && return 0
 
+if [[ "$ROBOT_REAL" != "true" ]]
+  then
+    read -r -p "This package ($target) should only be installed on the robot itself, not on your own PC. Are you sure to continue? [y/N] " response
+    response=${response,,}    # tolower
+    if [[ "$response" =~ ^(yes|y)$ ]]
+    then
+      :
+    else
+      return 1
+    fi
+fi
 
 ### Make sure remote launching works ###
 tue-install-cp ros-bash-and-run.sh ~/.ros-bash-and-run.sh
